revert json changes
