+++ src/relay/ir/doc.h
@@ -6,9 +6,9 @@
  * to you under the Apache License, Version 2.0 (the
  * "License"); you may not use this file except in compliance
  * with the License.  You may obtain a copy of the License at
- * 
+ *
  *   http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing,
  * software distributed under the License is distributed on an
  * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
@@ -48,8 +48,7 @@ struct TextNode : DocAtomNode {
 
   explicit TextNode(const std::string& str) : str(str) {
     if (str.find_first_of("\t\n") != str.npos) {
-      LOG(FATAL) << "text node: '" << str << "' should not has tab or newline.";
-      throw;
+      LOG(WARNING) << "text node: '" << str << "' should not has tab or newline.";
     }
   }
 };
